+++ src/utils/location_processor.h
@@ -5,7 +5,7 @@ namespace smp
 	class OnProcessLocationsNotify_InsertHandles : public process_locations_notify
 	{
 	public:
-		OnProcessLocationsNotify_InsertHandles(GUID g, UINT baseIdx, bool shouldSelect);
+		OnProcessLocationsNotify_InsertHandles(GUID g, size_t baseIdx, bool shouldSelect);
 
 		void on_completion(metadb_handle_list_cref items) final;
 		void on_aborted() final {}
